Revert "[scraper helper] Include support for timeout (#7703)"
This reverts commit c6b5e77f06f6d98091490927276111806a8682eb.
