--- .chloggen/msg_introduce-timeout-scrape.yaml
@@ -1,16 +1,0 @@
-# One of 'breaking', 'deprecation', 'new_component', 'enhancement', 'bug_fix'
-change_type: "enhancement"
-
-# The name of the component, or a single word describing the area of concern, (e.g. otlpreceiver)
-component: scraper helper
-
-# A brief description of the change.  Surround your text with quotes ("") if it needs to start with a backtick (`).
-note: Ensuring that scrapers must finished within their scrape interval
-
-# One or more tracking issues or pull requests related to the change
-issues: [7703]
-
-# (Optional) One or more lines of additional information to render under the primary note.
-# These lines will be padded with 2 spaces and then inserted directly into the document.
-# Use pipe (|) for multiline entries.
-subtext:
